Create README.md
Non intrusive load monitoring or NILM is a process of analysis of disaggregation of electric appliances connected to a power within a certain area. Dataset consists of power load data collected using current(mA) sensor attached to the wire of main power supply. Objective is to identify the individual device states(on/off/idle) for each combined load. This is a multilabel classification problem .Clustering techniques and multi label tree based models were used for the solving the problem.